Changed API end-point and method
I want to have a better way of defining the process. Signing up a user
should be separate from starting the service. For example, you sign up a
user with a `POST` to `/start` with `username`, `email`, and `accountname`,
and then at some point later you would `POST` to `/begin` with just the service
`username`. Perhaps the names should be reversed.