Added Nexmo endpoint
The endpoint is very simple and not fully complete. It currently always
assumes that there has been no git commit in the last 24 hours, and is
really just a complicated proxy for a simple `curl` call.

`curl --data 'number=15553334444&message=test+message' host:3000/test`
Will send the contents of `message` to `number`.